[software] Restructure systolic apps and runtime
