+++ software/apps/systolic/matmul/main.c
@@ -1,18 +1,6 @@
-// Copyright 2021 ETH Zurich and University of Bologna.
-//
+// Copyright 2022 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 // Author: Gua Hao Khov, ETH Zurich
 
@@ -21,8 +9,8 @@
 
 #include "alloc.h"
 #include "encoding.h"
-#include "kernel/queue.h"
-#include "kernel/systolic_a.h"
+#include "systolic/queue.h"
+#include "systolic/matmul.h"
 #include "printf.h"
 #include "runtime.h"
 #include "synchronization.h"
@@ -66,7 +54,7 @@ int main() {
   uint32_t num_cores = mempool_get_core_count();
 
   // Initialize synchronization variables
-  mempool_barrier_init(core_id, num_cores);
+  mempool_barrier_init(core_id);
 
   // Initialization
   mempool_init(core_id);
@@ -94,7 +82,7 @@ int main() {
   }
 
   // Wait for all cores
-  mempool_barrier(num_cores, num_cores * 4);
+  mempool_barrier(num_cores);
 
   // Assign grid position
   uint32_t col_idx = core_id % 4;
@@ -133,7 +121,7 @@ int main() {
   }
 
   // Wait for all cores
-  mempool_barrier(num_cores, num_cores * 4);
+  mempool_barrier(num_cores);
 
   // Print out benchmark
   if (core_id == 0) {
@@ -141,6 +129,6 @@ int main() {
   }
 
   // wait until all cores have finished
-  mempool_barrier(num_cores, num_cores * 4);
+  mempool_barrier(num_cores);
   return 0;
 }
